backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
websockets==12.0
pytest==7.4.4
pytest-asyncio==0.23.3